fix: read api_version from Docker version endpoint; i64 timeout
- docker_system_info now calls docker.version() via tokio::try_join!
  to populate the api_version field (was String::new())
- restart_container timeout uses i64 throughout the app layer,
  casting to isize only at the bollard boundary

+++ src-tauri/src/docker/containers.rs
@@ -248,7 +248,7 @@ impl<'a> ContainerOps<'a> {
 
     pub async fn restart(&self, id: &str, timeout: i64) -> Result<()> {
         let docker = self.client.get().await?;
-        let opts = RestartContainerOptions { t: timeout };
+        let opts = RestartContainerOptions { t: timeout as isize };
         docker
             .restart_container(id, Some(opts))
             .await
